+++ rwsplitter-jpa/src/main/java/com/elihullc/rwsplitter/jpa/hibernate/hikaricp/implementation/PropertiesFileMultiTenantConnectionProvider.java
@@ -3,6 +3,9 @@ package com.elihullc.rwsplitter.jpa.hibernate.hikaricp.implementation;
 import com.elihullc.rwsplitter.jpa.hibernate.SpringMultiTenantConnectionProvider;
 import com.elihullc.rwsplitter.jpa.hibernate.SpringTenantIdentifierResolver;
 
+import java.util.Map;
+import java.util.function.Supplier;
+
 public class PropertiesFileMultiTenantConnectionProvider
   extends SpringMultiTenantConnectionProvider<PropertiesFileDataSourceConnectionProvider> {
 
@@ -10,6 +13,16 @@ public class PropertiesFileMultiTenantConnectionProvider
 
     public PropertiesFileMultiTenantConnectionProvider(final SpringTenantIdentifierResolver tenantIdentifierResolver) {
         super(tenantIdentifierResolver);
+    }
+
+    public PropertiesFileMultiTenantConnectionProvider(SpringTenantIdentifierResolver tenantIdentifierResolver,
+      Supplier<Map<String, Boolean>> mapSupplier) {
+        super(tenantIdentifierResolver, mapSupplier);
+    }
+
+    public PropertiesFileMultiTenantConnectionProvider(SpringTenantIdentifierResolver tenantIdentifierResolver,
+      Map<String, Boolean> migratedTenants) {
+        super(tenantIdentifierResolver, migratedTenants);
     }
 
     @Override